+++ assignments/assignments_parser_test.go
@@ -1,4 +1,4 @@
-package web
+package assignments
 
 import (
 	"context"
