+++ web/bh.go
@@ -1,0 +1,6 @@
+package web
+
+type BaseHookOptions struct {
+	BaseURL string
+	Secret  string
+}
